requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
requests>=2.31.0
beautifulsoup4==4.12.2
python-multipart>=0.0.6
google-cloud-aiplatform>=1.70.0
vertexai>=1.71.0
google-generativeai>=0.3.2
python-dotenv>=1.0.0
lxml>=4.9.3
collections-extended>=2.0.2